requirements.txt
boto3==1.24.84
botocore==1.27.96
certifi==2025.8.3
fastapi==0.116.1
imbalanced_learn==0.12.4
imblearn==0.0
mypy_boto3_s3==1.24.76
neuro_mf==0.0.5
numpy==2.3.2
pandas==2.3.1
pymongo==4.2.0
python-dotenv==1.1.1
PyYAML==6.0.2
PyYAML==6.0.2
scikit_learn==1.7.1
scipy==1.16.1
setuptools==78.1.1
uvicorn==0.35.0
numpy
